Fixes circular dependecy issue
